requirements.txt
django>=1.8,<1.11
psycopg2
newspaper3k
django-cors-headers
tox
rest_framework

# TESTING
factory_boy
pytest
pytest-django
pytest-cov
pytest-flake8
pytest-mock
coverage
flake8

# debug toolbar
django-debug-toolbar